More refactor and coverage
