+++ SDKAndroid/src/test/kotlin/app/cybrid/sdkandroid/mock/Mocker.kt
@@ -1,0 +1,9 @@
+package app.cybrid.sdkandroid.mock
+
+import app.cybrid.sdkandroid.AppModule
+import io.mockk.mockkObject
+
+interface Mocker {
+
+    fun init() { mockkObject(AppModule) }
+}
